backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.2.1
email-validator==2.1.0.post1
python-dotenv==1.0.1
spacy==3.7.4
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1
httpx==0.26.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
bcrypt==4.1.2 